Modifique la forma de cambio para que habra el teclado numerico

+++ formas.py
@@ -1,14 +1,15 @@
 from flask_wtf import Form
 from wtforms import PasswordField, StringField, FloatField, FileField, SubmitField
 from wtforms.validators import DataRequired
+from wtforms.fields.html5 import DecimalField
 
 class Contrasena(Form):
     contrasena = PasswordField('Contraseña')
     login = SubmitField('Ingresar')
 
 class CambioForma(Form):
-    compra = FloatField('Compra', validators=[DataRequired()])
-    venta =  FloatField('Venta', validators=[DataRequired()])
+    compra = DecimalField('Compra', validators=[DataRequired()])
+    venta =  DecimalField('Venta', validators=[DataRequired()])
     enviar = SubmitField('Enviar')
 
 class FotoForma(Form):
